+++ pulsar-common/src/main/java/org/apache/pulsar/common/api/EncryptionContext.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pulsar.client.impl;
+package org.apache.pulsar.common.api;
 
 import java.util.Map;
 import java.util.Optional;
